chore: move workspace/doc/position into own files
