+++ pygls/protocol.py
@@ -105,7 +105,7 @@ from lsprotocol.types import (
     WorkDoneProgressCancelParams,
 )
 from pygls.uris import from_fs_path
-from pygls.workspace import Workspace
+from pygls.workspace.workspace import Workspace
 
 logger = logging.getLogger(__name__)
